requirements.txt
alembic==1.4.2
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
speedtest-cli==2.1.2
SQLAlchemy==1.3.19
urllib3==1.25.9
Werkzeug==1.0.1
